Improved the audio sys

+++ src/system_controller/system_controller/system_controller.py
@@ -506,7 +506,7 @@ class SystemController(Node):
         
         # angle: -1 = fully closed, +1 = fully open
         # Convert to motor position: jaw_max (closed) to jaw_min (open)
-        current_jaw_position = int(jaw_max - ((angle + 1)/2)*range)
+        current_jaw_position = jaw_max - ((angle + 1)/2)*range
         
         # Safety clamp to prevent motor damage
         current_jaw_position = max(jaw_min, min(jaw_max, current_jaw_position))
